update for length limits

+++ src/Domain/Entity/Construction/Joint.cs
@@ -16,8 +16,9 @@ namespace Prizm.Domain.Entity.Construction
         public virtual float DistanceFromKP { get; set; }
         public virtual DateTime? LoweringDate { get; set; }
         public virtual JointStatus Status { get; set; }
-        public virtual float GpsLatitude { get; set; }
-        public virtual float GpsLongitude { get; set; }
+        // for GpsLatitude and GpsLongitude double is required for зкщзук functioning os GPS mask (opportunity to enterring more than 9 characters)
+        public virtual double GpsLatitude { get; set; }
+        public virtual double GpsLongitude { get; set; }
         public virtual float GpsHeight { get; set; }
         public virtual PartData FirstElement { get; set; }
         public virtual PartData SecondElement { get; set; }
